Trekk funksjonell kode ut i egen funksjon for å gjenbrukes
